back port in .env. Right menu fix. comments . Alert fix
The backend port is now tied to the .env. A slight bug fixed applied to right menu when no graph exists. Some debug on file type control across the side bar. Some comments in Right Menu. UseEffect in side menu is not working as intended so its been commented out. Fixed the error alert system to align with the updated backend